Added Buying Ticket API

+++ BackEnd/park_management/src/main/java/park/management/com/vn/model/response/TicketDetailsResponse.java
@@ -10,6 +10,6 @@ public class TicketDetailsResponse {
     // Echoed from DB
     private BigDecimal price;
     private Integer quantity;
-    private Integer discount;
+    private BigDecimal discount; // updated from Integer
 
 }
